twister: move coverage classes out
Move coverage related classes out into own file.
